+++ models/unit_config.go
@@ -7,9 +7,9 @@ type UnitConfig struct {
 	Text    string   `xml:",chardata" json:"-" gqlgen:"text"`
 	Spear   struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -18,9 +18,9 @@ type UnitConfig struct {
 	} `xml:"spear" json:"spear" gqlgen:"spear"`
 	Sword struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -29,9 +29,9 @@ type UnitConfig struct {
 	} `xml:"sword" json:"sword" gqlgen:"sword"`
 	Axe struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -40,9 +40,9 @@ type UnitConfig struct {
 	} `xml:"axe" json:"axe" gqlgen:"axe"`
 	Archer struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -51,9 +51,9 @@ type UnitConfig struct {
 	} `xml:"archer" json:"archer" gqlgen:"archer"`
 	Spy struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -62,9 +62,9 @@ type UnitConfig struct {
 	} `xml:"spy" json:"spy" gqlgen:"spy"`
 	Light struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -73,9 +73,9 @@ type UnitConfig struct {
 	} `xml:"light" json:"light" gqlgen:"light"`
 	Marcher struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -84,9 +84,9 @@ type UnitConfig struct {
 	} `xml:"marcher" json:"marcher" gqlgen:"marcher"`
 	Heavy struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -95,9 +95,9 @@ type UnitConfig struct {
 	} `xml:"heavy" json:"heavy" gqlgen:"heavy"`
 	Ram struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -106,9 +106,9 @@ type UnitConfig struct {
 	} `xml:"ram" json:"ram" gqlgen:"ram"`
 	Catapult struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -117,9 +117,9 @@ type UnitConfig struct {
 	} `xml:"catapult" json:"catapult" gqlgen:"catapult"`
 	Knight struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -128,9 +128,9 @@ type UnitConfig struct {
 	} `xml:"knight" json:"knight" gqlgen:"knight"`
 	Snob struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
@@ -139,9 +139,9 @@ type UnitConfig struct {
 	} `xml:"snob" json:"snob" gqlgen:"snob"`
 	Militia struct {
 		Text           string  `xml:",chardata" json:"-" gqlgen:"text"`
-		BuildTime      float32 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
+		BuildTime      float64 `xml:"build_time" json:"build_time" gqlgen:"buildTime"`
 		Pop            int     `xml:"pop" json:"pop" gqlgen:"pop"`
-		Speed          float32 `xml:"speed" json:"speed" gqlgen:"speed"`
+		Speed          float64 `xml:"speed" json:"speed" gqlgen:"speed"`
 		Attack         int     `xml:"attack" json:"attack" gqlgen:"attack"`
 		Defense        int     `xml:"defense" json:"defense" gqlgen:"defense"`
 		DefenseCavalry int     `xml:"defense_cavalry" json:"defense_cavalry" gqlgen:"defenseCavalry"`
